Cleaned up compiler warnings, changed curl system calls to use libcurl

+++ cpp-client/apt-get.cpp
@@ -19,7 +19,7 @@ void AptGet::getAvailableUpdates(vector<updateInfo> & outList)
 
 	if(commandRetval == 0)
 	{
-		for(int i = 0; i < commandOutput.size(); i++)
+		for(size_t i = 0; i < commandOutput.size(); i++)
 		{
 			// example of one output line
 			// Inst libpam-modules [1.1.1-2ubuntu5] (1.1.1-2ubuntu5.3 Ubuntu:10.04/lucid-updates)
